Move test helpers to mantidimaging.core.testing
