+++ mantidimaging/core/filters/background_correction/test/background_correction_test.py
@@ -5,7 +5,7 @@ import unittest
 import numpy as np
 import numpy.testing as npt
 
-from mantidimaging.test import test_helper as th
+import mantidimaging.core.testing.unit_test_helper as th
 
 from mantidimaging.core.filters import background_correction
