+++ mantidimaging/core/filters/outliers/test/outliers_test.py
@@ -4,7 +4,7 @@ import unittest
 
 import numpy.testing as npt
 
-from mantidimaging.test import test_helper as th
+import mantidimaging.core.testing.unit_test_helper as th
 
 from mantidimaging.core.filters import outliers
