+++ mantidimaging/core/filters/ring_removal/test/ring_removal_test.py
@@ -4,8 +4,9 @@ import unittest
 
 import numpy.testing as npt
 
+import mantidimaging.core.testing.unit_test_helper as th
+
 from mantidimaging.core.utility.memory_usage import get_memory_usage_linux
-from mantidimaging.test import test_helper as th
 
 from mantidimaging.core.filters import ring_removal
