+++ mantidimaging/core/filters/roi_normalisation/test/roi_normalisation_test.py
@@ -5,7 +5,7 @@ import unittest
 import numpy as np
 import numpy.testing as npt
 
-from mantidimaging.test import test_helper as th
+import mantidimaging.core.testing.unit_test_helper as th
 
 from mantidimaging.core.filters import roi_normalisation
