+++ mantidimaging/core/parallel/test/exclusive_mem_test.py
@@ -4,8 +4,9 @@ import unittest
 
 import numpy.testing as npt
 
+import mantidimaging.core.testing.unit_test_helper as th
+
 from mantidimaging.core.utility.memory_usage import get_memory_usage_linux
-from mantidimaging.test import test_helper as th
 
 from mantidimaging.core.parallel import exclusive_mem as esm
