+++ mantidimaging/core/testing/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import (absolute_import, division, print_function)
+
+from .file_outputting_test_case import FileOutputtingTestCase  # noqa: F401
+
+del absolute_import, division, print_function  # noqa:F821
