+++ mantidimaging/core/testing/file_outputting_test_case.py
